+++ exporters/ostream/test/ostream_metric_test.cc
@@ -23,7 +23,7 @@ namespace exportermetrics = opentelemetry::exporter::metrics;
 TEST(OStreamMetricsExporter, Shutdown)
 {
   auto exporter =
-      std::unique_ptr<metric_sdk::MetricExporter>(new exportermetrics::OStreamMetricExporter);
+      std::unique_ptr<metric_sdk::PushMetricExporter>(new exportermetrics::OStreamMetricExporter);
   ASSERT_TRUE(exporter->Shutdown());
   auto result = exporter->Export(metric_sdk::ResourceMetrics{});
   EXPECT_EQ(result, opentelemetry::sdk::common::ExportResult::kFailure);
@@ -32,7 +32,7 @@ TEST(OStreamMetricsExporter, Shutdown)
 TEST(OStreamMetricsExporter, ExportSumPointData)
 {
   auto exporter =
-      std::unique_ptr<metric_sdk::MetricExporter>(new exportermetrics::OStreamMetricExporter);
+      std::unique_ptr<metric_sdk::PushMetricExporter>(new exportermetrics::OStreamMetricExporter);
 
   metric_sdk::SumPointData sum_point_data{};
   sum_point_data.value_ = 10.0;
@@ -95,7 +95,7 @@ TEST(OStreamMetricsExporter, ExportSumPointData)
 TEST(OStreamMetricsExporter, ExportHistogramPointData)
 {
   auto exporter =
-      std::unique_ptr<metric_sdk::MetricExporter>(new exportermetrics::OStreamMetricExporter);
+      std::unique_ptr<metric_sdk::PushMetricExporter>(new exportermetrics::OStreamMetricExporter);
 
   metric_sdk::HistogramPointData histogram_point_data{};
   histogram_point_data.boundaries_ = std::list<double>{10.1, 20.2, 30.2};
@@ -177,7 +177,7 @@ TEST(OStreamMetricsExporter, ExportHistogramPointData)
 TEST(OStreamMetricsExporter, ExportLastValuePointData)
 {
   auto exporter =
-      std::unique_ptr<metric_sdk::MetricExporter>(new exportermetrics::OStreamMetricExporter);
+      std::unique_ptr<metric_sdk::PushMetricExporter>(new exportermetrics::OStreamMetricExporter);
 
   metric_sdk::ResourceMetrics data;
   auto resource = opentelemetry::sdk::resource::Resource::Create(
@@ -246,7 +246,7 @@ TEST(OStreamMetricsExporter, ExportLastValuePointData)
 TEST(OStreamMetricsExporter, ExportDropPointData)
 {
   auto exporter =
-      std::unique_ptr<metric_sdk::MetricExporter>(new exportermetrics::OStreamMetricExporter);
+      std::unique_ptr<metric_sdk::PushMetricExporter>(new exportermetrics::OStreamMetricExporter);
 
   metric_sdk::ResourceMetrics data;
   auto resource = opentelemetry::sdk::resource::Resource::Create(