+++ exporters/otlp/test/otlp_http_metric_exporter_factory_test.cc
@@ -35,7 +35,7 @@ TEST(OtlpHttpMetricExporterFactory, BuildTest)
   OtlpHttpMetricExporterOptions opts;
   opts.url = "localhost:45454";
 
-  std::unique_ptr<opentelemetry::sdk::metrics::MetricExporter> exporter =
+  std::unique_ptr<opentelemetry::sdk::metrics::PushMetricExporter> exporter =
       OtlpHttpMetricExporterFactory::Create(opts);
 
   EXPECT_TRUE(exporter != nullptr);